move test data to JsonConverterTestBase
