+++ hudi-common/src/test/java/org/apache/hudi/avro/TestMercifulJsonConverter.java
@@ -33,7 +33,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -47,16 +46,15 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
-import java.util.UUID;
 import java.util.stream.Stream;
 
 import static org.apache.hudi.common.testutils.HoodieTestDataGenerator.TRIP_ENCODED_DECIMAL_SCHEMA;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class TestMercifulJsonConverter {
+public class TestMercifulJsonConverter extends JsonConverterTestBase {
   private static final ObjectMapper MAPPER = new ObjectMapper();
-  private static final MercifulJsonConverter CONVERTER = new MercifulJsonConverter(true,"__");
+  private static final MercifulJsonConverter CONVERTER = new MercifulJsonConverter(true, "__");
 
   @Test
   public void basicConversion() throws IOException {
@@ -79,11 +77,7 @@ public class TestMercifulJsonConverter {
   }
 
   @ParameterizedTest
-  @ValueSource(strings = {
-      "{\"first\":\"John\",\"last\":\"Smith\"}",
-      "[{\"first\":\"John\",\"last\":\"Smith\"}]",
-      "{\"first\":\"John\",\"last\":\"Smith\",\"suffix\":3}",
-  })
+  @MethodSource("nestedJsonAsString")
   void nestedJsonAsString(String nameInput) throws IOException {
     Schema simpleSchema = SchemaTestUtil.getSimpleSchema();
     String json = String.format("{\"name\": %s, \"favorite_number\": 1337, \"favorite_color\": 10}", nameInput);
@@ -96,7 +90,6 @@ public class TestMercifulJsonConverter {
     assertEquals(rec, CONVERTER.convert(json, simpleSchema));
   }
 
-  private static final String DECIMAL_AVRO_FILE_INVALID_PATH = "/decimal-logical-type-invalid.avsc";
   private static final String DECIMAL_AVRO_FILE_PATH = "/decimal-logical-type.avsc";
   private static final String DECIMAL_ZERO_SCALE_AVRO_FILE_PATH = "/decimal-logical-type-zero-scale.avsc";
 
@@ -126,7 +119,6 @@ public class TestMercifulJsonConverter {
     assertEquals(record, real);
   }
 
-  private static final String DECIMAL_FIXED_AVRO_FILE_PATH = "/decimal-logical-type-fixed-type.avsc";
   /**
    * Covered case:
    * Avro Logical Type: Decimal
@@ -149,27 +141,6 @@ public class TestMercifulJsonConverter {
     assertThrows(HoodieJsonToAvroConversionException.class, () -> {
       CONVERTER.convert(json, schema);
     });
-  }
-
-  static Stream<Object> decimalBadCases() {
-    return Stream.of(
-        // Invalid schema definition.
-        Arguments.of(DECIMAL_AVRO_FILE_INVALID_PATH, "123.45", null),
-        // Schema set precision as 5, input overwhelmed the precision.
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "123333.45", null),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, null, 123333.45),
-        // Schema precision set to 5, scale set to 2, so there is only 3 digit to accommodate integer part.
-        // As we do not do rounding, any input with more than 3 digit integer would fail.
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "1233", null),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, null, 1233D),
-        // Schema set scale as 2, input overwhelmed the scale.
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "0.222", null),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, null, 0.222),
-        // Invalid string which cannot be parsed as number.
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "", null),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "NotAValidString", null),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "-", null)
-    );
   }
 
   /**
@@ -225,46 +196,6 @@ public class TestMercifulJsonConverter {
     assertEquals(record, real);
   }
 
-  static Stream<Object> decimalGoodCases() {
-    return Stream.of(
-        // The schema all set precision as 5, scale as 2.
-        // Test dimension: Schema file, Ground truth, string input, number input, fixed byte array input.
-        // Test some random numbers.
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "123.45", "123.45", null, false),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "123.45", null, 123.45, false),
-        // Test MIN/MAX allowed by the schema.
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "-999.99", "-999.99", null, false),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "999.99", null, 999.99, false),
-        // Test 0.
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "0", null, 0D, false),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "0", "0", null, false),
-        Arguments.of(DECIMAL_AVRO_FILE_PATH, "0", "000.00", null, false),
-        // Same set of coverage over schame using byte/fixed type.
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "123.45", "123.45", null, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "123.45", null, 123.45, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "-999.99", "-999.99", null, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "999.99", null, 999.99, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "999", null, 999, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "999", null, 999L, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "999", null, (short) 999, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "100", null, (byte) 100, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "0", null, 0D, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "0", null, 0, false),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "0", "0", null, true),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "0", "000.00", null, true),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "123.45", null, null, true),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "123.45", null, 123.45, true),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "-999.99", null, null, true),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "999.99", null, 999.99, true),
-        Arguments.of(DECIMAL_FIXED_AVRO_FILE_PATH, "0", null, null, true),
-        Arguments.of(DECIMAL_ZERO_SCALE_AVRO_FILE_PATH, "12345", "12345.0", null, false),
-        Arguments.of(DECIMAL_ZERO_SCALE_AVRO_FILE_PATH, "12345", null, 12345.0, false),
-        Arguments.of(DECIMAL_ZERO_SCALE_AVRO_FILE_PATH, "12345", null, 12345, false),
-        Arguments.of(DECIMAL_ZERO_SCALE_AVRO_FILE_PATH, "1230", null, 1.23e+3, false),
-        Arguments.of(DECIMAL_ZERO_SCALE_AVRO_FILE_PATH, "1230", "1.23e+3", null, false)
-    );
-  }
-
   // tests cases where decimals with fraction `.0` can be interpreted as having scale > 0
   @ParameterizedTest
   @MethodSource("zeroScaleDecimalCases")
@@ -286,35 +217,6 @@ public class TestMercifulJsonConverter {
     }
   }
 
-  static Stream<Object> zeroScaleDecimalCases() {
-    return Stream.of(
-        // Input value in JSON, expected decimal, whether conversion should be successful
-        // Values that can be converted
-        Arguments.of("0.0", "0", true),
-        Arguments.of("20.0", "20", true),
-        Arguments.of("320", "320", true),
-        Arguments.of("320.00", "320", true),
-        Arguments.of("-1320.00", "-1320", true),
-        Arguments.of("1520423524459", "1520423524459", true),
-        Arguments.of("1520423524459.0", "1520423524459", true),
-        Arguments.of("1000000000000000.0", "1000000000000000", true),
-        // Values that are big enough and out of range of int or long types
-        // Note that we can have at most 17 significant decimal digits in double values
-        Arguments.of("1.2684037455962608e+16", "12684037455962608", true),
-        Arguments.of("4.0100001e+16", "40100001000000000", true),
-        Arguments.of("3.52838e+17", "352838000000000000", true),
-        Arguments.of("9223372036853999600.0000", "9223372036853999600", true),
-        Arguments.of("999998887654321000000000000000.0000", "999998887654321000000000000000", true),
-        Arguments.of("-999998887654321000000000000000.0000", "-999998887654321000000000000000", true),
-        // Values covering high precision decimals that lose precision when converting to a double
-        Arguments.of("3.781239258857277e+16", "37812392588572770", true),
-        Arguments.of("1.6585135379127473e+18", "1658513537912747300", true),
-        // Values that should not be converted
-        Arguments.of("0.0001", null, false),
-        Arguments.of("300.9999", null, false),
-        Arguments.of("1928943043.0001", null, false)
-    );
-  }
 
   private static final String DURATION_AVRO_FILE_PATH = "/duration-logical-type.avsc";
   /**
@@ -349,29 +251,15 @@ public class TestMercifulJsonConverter {
     assertEquals(durationRecord, real);
   }
 
-  static Stream<Object> durationGoodCases() {
-    return Stream.of(
-        // Normal inputs.
-        Arguments.of(1, 2, 3),
-        // Negative int would be interpreted as some unsigned int by Avro. They all 4-byte.
-        Arguments.of(-1, -2, -3),
-        // Signed -1 interpreted to unsigned would be unsigned MAX
-        Arguments.of(-1, -1, -1),
-        // Other special edge cases.
-        Arguments.of(0, 0, 0),
-        Arguments.of(Integer.MAX_VALUE, Integer.MAX_VALUE, Integer.MAX_VALUE),
-        Arguments.of(Integer.MIN_VALUE, Integer.MIN_VALUE, Integer.MIN_VALUE)
-    );
-  }
-
-  @Test
-  void durationLogicalTypeBadTest() throws IOException {
+  @ParameterizedTest
+  @MethodSource("durationBadCases")
+  void durationLogicalTypeBadTest(Long months, Long days, Long millis) throws IOException {
     // As duration uses 12 byte fixed type to store 3 unsigned int numbers, Long.MAX would cause overflow.
     // Verify it is gracefully handled.
     List<Long> num = new ArrayList<>();
-    num.add(Long.MAX_VALUE);
-    num.add(Long.MAX_VALUE);
-    num.add(Long.MAX_VALUE);
+    num.add(months);
+    num.add(days);
+    num.add(millis);
 
     Map<String, Object> data = new HashMap<>();
     data.put("duration", num);
@@ -406,17 +294,6 @@ public class TestMercifulJsonConverter {
     String json = MAPPER.writeValueAsString(data);
     GenericRecord real = CONVERTER.convert(json, schema);
     assertEquals(record, real);
-  }
-
-  static Stream<Object> dataProvider() {
-    return Stream.of(
-        Arguments.of(18506, 18506), // epochDays
-        Arguments.of(18506, "2020-09-01"),  // dateString
-        Arguments.of(7323356, "+22020-09-01"),  // dateString
-        Arguments.of(18506, "18506"),  // epochDaysString
-        Arguments.of(Integer.MAX_VALUE, Integer.toString(Integer.MAX_VALUE)),
-        Arguments.of(Integer.MIN_VALUE, Integer.toString(Integer.MIN_VALUE))
-    );
   }
 
   /**
@@ -469,83 +346,6 @@ public class TestMercifulJsonConverter {
     assertEquals(record, real);
   }
 
-  static Stream<Object> localTimestampGoodCaseProvider() {
-    return Stream.of(
-        // Test cases with 'T' as the separator
-        Arguments.of(
-            (long)(1715644416 * 1e6 + 4000000 / 1e3), // Num of micro sec since unix epoch
-            "2024-05-13T23:53:36.004", // Timestamp equivalence
-            "2024-05-13T23:53:36.004"),
-        Arguments.of(
-            (long)(1715644416 * 1e6), // Num of micro sec since unix epoch
-            "2024-05-13T23:53:36", // Timestamp equivalence
-            "2024-05-13T23:53:36"),
-        // Test cases with ' ' as the separator
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 4000000 / 1e3), // Num of micro sec since unix epoch
-            "2024-05-13 23:53:36.004", // Timestamp equivalence
-            "2024-05-13 23:53:36.004"),
-        Arguments.of(
-            (long) (1715644416 * 1e6), // Num of micro sec since unix epoch
-            "2024-05-13 23:53:36", // Timestamp equivalence
-            "2024-05-13 23:53:36"),
-        Arguments.of(
-            2024L, "2", "2024"),
-        Arguments.of(
-            (long)(1715644416 * 1e6 + 4000000 / 1e3),
-            (long)(1715644416 * 1e3 + 4000000 / 1e6),
-            (long)(1715644416 * 1e6 + 4000000 / 1e3)),
-        Arguments.of(
-            (long)(1715644416 * 1e6 + 4000000 / 1e3),
-            (long)(1715644416 * 1e3 + 4000000 / 1e6),
-            Long.toString((long)(1715644416 * 1e6 + 4000000 / 1e3))),
-        // Test higher precision that only micro sec unit can capture.
-        Arguments.of(
-            (long)(1715644416 * 1e6 + 4000000 / 1e6),
-            "2024-05-13T23:53:36.000", // Timestamp equivalence
-            "2024-05-13T23:53:36.000004"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 4000000 / 1e6),
-            "2024-05-13 23:53:36.000", // Timestamp equivalence
-            "2024-05-13 23:53:36.000004"),
-        // Test full range of time
-        Arguments.of(
-            0L,
-            "1970-01-01T00:00:00.000", // Timestamp equivalence
-            "1970-01-01T00:00:00.000000"),
-        Arguments.of(
-            0L,
-            "1970-01-01 00:00:00.000", // Timestamp equivalence
-            "1970-01-01 00:00:00.000000"),
-        Arguments.of(
-            Long.MAX_VALUE,
-            "+294247-01-10T04:00:54.775", // Timestamp in far future must be prefixed with '+'
-            "+294247-01-10T04:00:54.775807"),
-        Arguments.of(
-            Long.MAX_VALUE,
-            "+294247-01-10 04:00:54.775", // Timestamp in far future must be prefixed with '+'
-            "+294247-01-10 04:00:54.775807"),
-        Arguments.of(
-            0L, 0L, 0L),
-        Arguments.of(
-            -1L * 1000, -1L, -1L * 1000),
-        Arguments.of(
-            Long.MIN_VALUE, Long.MIN_VALUE / 1000, Long.MIN_VALUE),
-        Arguments.of(
-            Long.MAX_VALUE, Long.MAX_VALUE / 1000, Long.MAX_VALUE),
-        Arguments.of(
-            -62167219200000000L, "0000-01-01T00:00:00.00000", "0000-01-01T00:00:00.00000"),
-            Arguments.of(
-                -62167219200000000L, -62167219200000000L / 1000, -62167219200000000L),
-        Arguments.of(
-            -62167219200000000L, "0000-01-01 00:00:00.00000", "0000-01-01 00:00:00.00000"),
-        Arguments.of(
-            -62167219200000000L, -62167219200000000L / 1000, -62167219200000000L)
-    );
-  }
-
-  private static final String LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH = "/local-timestamp-millis-logical-type.avsc";
-  private static final String LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH = "/local-timestamp-micros-logical-type.avsc";
   @ParameterizedTest
   @MethodSource("localTimestampBadCaseProvider")
   void localTimestampLogicalTypeBadTest(
@@ -559,45 +359,6 @@ public class TestMercifulJsonConverter {
     assertThrows(HoodieJsonToAvroConversionException.class, () -> {
       CONVERTER.convert(json, schema);
     });
-  }
-
-  static Stream<Object> localTimestampBadCaseProvider() {
-    return Stream.of(
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-05-1323:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-05-1T23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-05-1 23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-0-13T23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-0-13 23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "20242-05-13T23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "20242-05-13 23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "202-05-13T23:53:36.0000000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "202-05-13 23:53:36.0000000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "202-05-13T23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "202-05-13 23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-05-13T23:53:36.000Z"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-05-13 23:53:36.000Z"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024-05-1323:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024-05-1T23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024-05-1 23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024-0-13T23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024-0-13 23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "20242-05-13T23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "20242-05-13 23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "202-05-13T23:53:36.0000000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "202-05-13 23:53:36.0000000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "202-05-13T23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "202-05-13 23:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024-05-13T23:53:36.000Z"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024-05-13 23:53:36.000Z"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "Not a timestamp at all!"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024 05 13T23:00"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024 05 13 23:00"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2024-05"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2011-12-03T10:15:30+01:00"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2011-12-03 10:15:30+01:00"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2011-12-03T10:15:30[Europe/ Paris]"),
-        Arguments.of(LOCAL_TIMESTAMP_MICRO_AVRO_FILE_PATH, "2011-12-03 10:15:30[Europe/ Paris]")
-    );
   }
 
   private static final String TIMESTAMP_AVRO_FILE_PATH = "/timestamp-logical-type2.avsc";
@@ -630,131 +391,6 @@ public class TestMercifulJsonConverter {
     assertEquals(record, real);
   }
 
-  static Stream<Object> timestampGoodCaseProvider() {
-    return Stream.of(
-        Arguments.of(
-            (long)(1715644416 * 1e6 + 4000000 / 1e3), // Num of micro sec since unix epoch
-            "2024-05-13T23:53:36.004Z", // Timestamp equivalence
-            "2024-05-13T23:53:36.004Z"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 4000000 / 1e3), // Num of micro sec since unix epoch
-            "2024-05-13 23:53:36.004Z", // Timestamp equivalence
-            "2024-05-13 23:53:36.004Z"),
-        Arguments.of(
-            (long) (1715644416 * 1e6), // Num of micro sec since unix epoch
-            "2024-05-13T23:53:36Z", // Timestamp equivalence
-            "2024-05-13T23:53:36Z"),
-        Arguments.of(
-            (long) (1715644416 * 1e6), // Num of micro sec since unix epoch
-            "2024-05-13 23:53:36Z", // Timestamp equivalence
-            "2024-05-13 23:53:36Z"),
-        // Test timestamps with no zone offset
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 4000000 / 1e3),
-            "2024-05-13T23:53:36.004",
-            "2024-05-13T23:53:36.004"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 4000000 / 1e3),
-            "2024-05-13 23:53:36.004",
-            "2024-05-13 23:53:36.004"),
-        // Test timestamps with different zone offsets
-        Arguments.of(
-            (long) (1715644416 * 1e6 - 2 * 3600 * 1e6),
-            "2024-05-13T23:53:36+02:00",
-            "2024-05-13T23:53:36+02:00"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 - 2 * 3600 * 1e6),
-            "2024-05-13 23:53:36+02:00",
-            "2024-05-13 23:53:36+02:00"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 4000000 / 1e3),
-            "2024-05-13T23:53:36.004+00:00",
-            "2024-05-13T23:53:36.004+00:00"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 4000000 / 1e3),
-            "2024-05-13 23:53:36.004+00:00",
-            "2024-05-13 23:53:36.004+00:00"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 - 3 * 3600 * 1e6 + 4000000 / 1e3),
-            "2024-05-13T23:53:36.004+03:00",
-            "2024-05-13T23:53:36.004+03:00"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 - 3 * 3600 * 1e6 + 4000000 / 1e3),
-            "2024-05-13 23:53:36.004+03:00",
-            "2024-05-13 23:53:36.004+03:00"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 6 * 3600 * 1e6 + 4000000 / 1e3),
-            "2024-05-13T23:53:36.004-06:00",
-            "2024-05-13T23:53:36.004-06:00"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 6 * 3600 * 1e6 + 4000000 / 1e3),
-            "2024-05-13 23:53:36.004-06:00",
-            "2024-05-13 23:53:36.004-06:00"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + (8 * 3600 + 1800) * 1e6 + 4000000 / 1e3),
-            "2024-05-13T23:53:36.004-08:30",
-            "2024-05-13T23:53:36.004-08:30"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + (8 * 3600 + 1800) * 1e6 + 4000000 / 1e3),
-            "2024-05-13 23:53:36.004-08:30",
-            "2024-05-13 23:53:36.004-08:30"),
-        Arguments.of(
-            2024L, "2", "2024"),
-        Arguments.of(
-            (long)(1715644416 * 1e6 + 4000000 / 1e3),
-            (long)(1715644416 * 1e3 + 4000000 / 1e6),
-            (long)(1715644416 * 1e6 + 4000000 / 1e3)),
-        Arguments.of(
-            (long)(1715644416 * 1e6 + 4000000 / 1e3),
-            (long)(1715644416 * 1e3 + 4000000 / 1e6),
-            Long.toString((long)(1715644416 * 1e6 + 4000000 / 1e3))),
-        // Test higher precision that only micro sec unit can capture.
-        Arguments.of(
-            (long)(1715644416 * 1e6 + 4000000 / 1e6),
-            "2024-05-13T23:53:36.000Z", // Timestamp equivalence
-            "2024-05-13T23:53:36.000004Z"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + 4000000 / 1e6),
-            "2024-05-13 23:53:36.000Z",
-            "2024-05-13 23:53:36.000004Z"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + (2 * 3600 + 1800) * 1e6 + 4000000 / 1e6),
-            "2024-05-13T23:53:36.000-02:30",
-            "2024-05-13T23:53:36.000004-02:30"),
-        Arguments.of(
-            (long) (1715644416 * 1e6 + (2 * 3600 + 1800) * 1e6 + 4000000 / 1e6),
-            "2024-05-13 23:53:36.000-02:30",
-            "2024-05-13 23:53:36.000004-02:30"),
-        // Test full range of time
-        Arguments.of(
-            0L,
-            "1970-01-01T00:00:00.000Z", // Timestamp equivalence
-            "1970-01-01T00:00:00.000000Z"),
-        Arguments.of(
-            (long) (-3600 * 1e6),
-            "1970-01-01T00:00:00.000+01:00",
-            "1970-01-01T00:00:00.000000+01:00"),
-        // The test case leads to long overflow due to how java calculate duration between 2 timestamps
-        // Arguments.of(
-        //  Long.MAX_VALUE,
-        //  "+294247-01-10T04:00:54.775Z", // Timestamp in far future must be prefixed with '+'
-        //  "+294247-01-10T04:00:54.775807Z"),
-        Arguments.of(
-            0L, 0L, 0L),
-        Arguments.of(
-            -1L * 1000, -1L, -1L * 1000),
-        Arguments.of(
-            Long.MIN_VALUE, Long.MIN_VALUE / 1000, Long.MIN_VALUE),
-        Arguments.of(
-            Long.MAX_VALUE, Long.MAX_VALUE / 1000, Long.MAX_VALUE),
-        // The test case leads to long overflow due to how java calculate duration between 2 timestamps
-        // Arguments.of(
-        //  -62167219200000000L, "0000-01-01T00:00:00.00000Z", "0000-01-01T00:00:00.00000Z"),
-        Arguments.of(
-        -62167219200000000L, -62167219200000000L / 1000, -62167219200000000L)
-    );
-  }
-
   @ParameterizedTest
   @MethodSource("timestampBadCaseProvider")
   void timestampLogicalTypeBadTest(Object input) throws IOException {
@@ -768,14 +404,6 @@ public class TestMercifulJsonConverter {
     assertThrows(HoodieJsonToAvroConversionException.class, () -> {
       CONVERTER.convert(json, schema);
     });
-  }
-
-  static Stream<Object> timestampBadCaseProvider() {
-    return Stream.of(
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-05-1323:53:36.000"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "2024-05-1323:53:36.000 UTC"),
-        Arguments.of(LOCAL_TIMESTAMP_MILLI_AVRO_FILE_PATH, "Tue, 3 Jun 2008 11:05:30 GMT")
-    );
   }
 
   private static final String TIME_AVRO_FILE_PATH = "/time-logical-type.avsc";
@@ -807,35 +435,6 @@ public class TestMercifulJsonConverter {
     assertEquals(record, real);
   }
 
-  static Stream<Object> timeGoodCaseProvider() {
-    return Stream.of(
-        // 12 hours and 30 minutes in milliseconds / microseconds
-        Arguments.of((long)4.5e10, (int)4.5e7, (long)4.5e10),
-        // 12 hours and 30 minutes in milliseconds / microseconds as string
-        Arguments.of((long)4.5e10, Integer.toString((int)4.5e7), Long.toString((long)4.5e10)),
-        // 12 hours and 30 minutes
-        Arguments.of((long)4.5e10, "12:30:00", "12:30:00"),
-        Arguments.of(
-            (long)(4.5e10 + 1e3), // 12 hours, 30 minutes and 0.001 seconds in microseconds
-            "12:30:00.001", // 12 hours, 30 minutes and 0.001 seconds
-            "12:30:00.001" // 12 hours, 30 minutes and 0.001 seconds
-            ),
-        // Test value ranges
-        Arguments.of(
-            0L,
-            "00:00:00.000",
-            "00:00:00.00000"
-            ),
-        Arguments.of(
-            86399999990L,
-            "23:59:59.999",
-            "23:59:59.99999"
-            ),
-        Arguments.of((long)Integer.MAX_VALUE, Integer.MAX_VALUE / 1000, (long)Integer.MAX_VALUE),
-        Arguments.of((long)Integer.MIN_VALUE, Integer.MIN_VALUE / 1000, (long)Integer.MIN_VALUE)
-        );
-  }
-
   @ParameterizedTest
   @MethodSource("timeBadCaseProvider")
   void timeLogicalTypeBadCaseTest(Object timeMilli, Object timeMicro) throws IOException {
@@ -850,12 +449,6 @@ public class TestMercifulJsonConverter {
     assertThrows(HoodieJsonToAvroConversionException.class, () -> {
       CONVERTER.convert(json, schema);
     });
-  }
-
-  static Stream<Object> timeBadCaseProvider() {
-    return Stream.of(
-        Arguments.of(0L, "00:0", "00:0")
-    );
   }
 
   private static final String UUID_AVRO_FILE_PATH = "/uuid-logical-type.avsc";
@@ -881,14 +474,28 @@ public class TestMercifulJsonConverter {
     assertEquals(record, real);
   }
 
-  static Stream<Object> uuidDimension() {
-    return Stream.of(
-      // Normal UUID
-      UUID.randomUUID().toString(),
-      // Arbitrary string will also pass as neither Avro library nor json convertor validate the string content.
-      "",
-      "NotAnUUID"
-    );
+  @ParameterizedTest
+  @MethodSource("nestedRecord")
+  void nestedRecordTest(String contactInput, boolean isString) {
+    String nestedSchemaStr = "{\"type\":\"record\",\"name\":\"User\",\"fields\":[{\"name\":\"name\",\"type\":\"string\"}," +
+        "{\"name\":\"contact\",\"type\":{\"type\":\"record\",\"name\":\"Contact\",\"fields\":[{\"name\":\"email\",\"type\":\"string\"}]}}]}";
+    String json = isString ? String.format("{\"name\":\"Jane Smith\",\"contact\":{\"email\":\"%s\"}}", contactInput)
+        : String.format("{\"name\":\"Jane Smith\",\"contact\":{\"email\":%s}}", contactInput);
+    Schema nestedSchema = new Schema.Parser().parse(nestedSchemaStr);
+    GenericRecord userRecord = new GenericData.Record(nestedSchema);
+
+    // Create the nested record for Contact
+    Schema contactSchema = nestedSchema.getField("contact").schema();
+    GenericRecord contactRecord = new GenericData.Record(contactSchema);
+
+    // Set the email field in the nested Contact record
+    contactRecord.put("email", contactInput);
+
+    // Set the fields in the outer User record
+    userRecord.put("name", "Jane Smith");
+    userRecord.put("contact", contactRecord);
+
+    assertEquals(userRecord, CONVERTER.convert(json, nestedSchema));
   }
 
   @Test
